Make a base class for testable callbacks
Have TestableNetworkCallback adopt it, as a first client. It's also by
far the most complex.

Test: TH
Flag: EXEMPT test-only

+++ tests/cts/hostside/app/src/com/android/cts/net/hostside/VpnTest.java
@@ -198,7 +198,7 @@ public class VpnTest {
     private static final String PRIVATE_DNS_MODE_PROVIDER_HOSTNAME = "hostname";
     private static final String PRIVATE_DNS_MODE_OPPORTUNISTIC = "opportunistic";
     private static final String PRIVATE_DNS_SPECIFIER_SETTING = "private_dns_specifier";
-    private static final int NETWORK_CALLBACK_TIMEOUT_MS = 30_000;
+    private static final long NETWORK_CALLBACK_TIMEOUT_MS = 30_000;
 
     private static final LinkAddress TEST_IP4_DST_ADDR = new LinkAddress("198.51.100.1/24");
     private static final LinkAddress TEST_IP4_SRC_ADDR = new LinkAddress("198.51.100.2/24");
@@ -1945,18 +1945,18 @@ public class VpnTest {
 
             expectVpnNetwork(defaultNetworkCallback);
         }, /* cleanup */ () -> {
-                stopVpn();
-                defaultNetworkCallback.eventuallyExpect(Event.LOST);
-            }, /* cleanup */ () -> {
-                runWithShellPermissionIdentity(() -> {
-                    mCM.setVpnDefaultForUids(session, new ArraySet<>());
-                }, NETWORK_SETTINGS);
-                // The default network of the app will be changed back to wifi when the VPN network
-                // preference feature is disabled.
-                defaultNetworkCallback.eventuallyExpect(Event.AVAILABLE,
-                        NETWORK_CALLBACK_TIMEOUT_MS,
-                        entry -> defaultNetwork.equals(entry.getNetwork()));
-            });
+            stopVpn();
+            defaultNetworkCallback.eventuallyExpect(Event.LOST);
+        }, /* cleanup */ () -> {
+            runWithShellPermissionIdentity(() -> {
+                mCM.setVpnDefaultForUids(session, new ArraySet<>());
+            }, NETWORK_SETTINGS);
+            // The default network of the app will be changed back to wifi when the VPN network
+            // preference feature is disabled.
+            defaultNetworkCallback.eventuallyExpect(Event.AVAILABLE,
+                    NETWORK_CALLBACK_TIMEOUT_MS,
+                    entry -> defaultNetwork.equals(entry.getNetwork()));
+        });
     }
 
     /**
